requirements.txt
Faab~=0.1.3
flasgger~=0.9.7.1
flask~=3.0.0
werkzeug~=3.0.1
Pillow~=10.1.0
Flask-JWT~=0.3.2
SQLAlchemy~=2.0.23
setuptools~=68.2.0
ffmpeg-python~=0.2.0